+++ src/main/java/com/tesa/hospitalerd/controller/PatientController.java
@@ -1,7 +1,7 @@
 package com.tesa.hospitalerd.controller;
 
-import com.tesa.hospitalerd.model.dto.PatientRequest;
-import com.tesa.hospitalerd.service.PatientService;
+import com.tesa.hospitalerd.model.request.PatientRequest;
+import com.tesa.hospitalerd.service.interfaces.PatientService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
