--- src/main/java/com/tesa/hospitalerd/model/dto/Response.java
@@ -1,15 +1,0 @@
-package com.tesa.hospitalerd.model.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Response {
-    private String responseCode;
-    private String responseMessage;
-}
