+++ src/main/java/com/tesa/hospitalerd/service/implementation/PatientServiceImpl.java
@@ -1,9 +1,10 @@
-package com.tesa.hospitalerd.service;
+package com.tesa.hospitalerd.service.implementation;
 
 import com.google.gson.Gson;
-import com.tesa.hospitalerd.model.dto.PatientRequest;
+import com.tesa.hospitalerd.model.request.PatientRequest;
 import com.tesa.hospitalerd.model.entity.Patient;
-import com.tesa.hospitalerd.repository.interfaces.PatientRepository;
+import com.tesa.hospitalerd.repository.database.interfaces.PatientRepository;
+import com.tesa.hospitalerd.service.interfaces.PatientService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
-public class PatientServiceImpl implements PatientService{
+public class PatientServiceImpl implements PatientService {
 
     private final PatientRepository patientRepository;
     Gson gson = new Gson();
