+++ src/main/java/com/tesa/hospitalerd/service/implementation/StaffServiceImpl.java
@@ -1,9 +1,10 @@
-package com.tesa.hospitalerd.service;
+package com.tesa.hospitalerd.service.implementation;
 
 import com.google.gson.Gson;
-import com.tesa.hospitalerd.model.dto.StaffRequest;
+import com.tesa.hospitalerd.model.request.StaffRequest;
 import com.tesa.hospitalerd.model.entity.Staff;
-import com.tesa.hospitalerd.repository.interfaces.StaffRepository;
+import com.tesa.hospitalerd.repository.database.interfaces.StaffRepository;
+import com.tesa.hospitalerd.service.interfaces.StaffService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
-public class StaffServiceImpl implements StaffService{
+public class StaffServiceImpl implements StaffService {
 
     private final StaffRepository staffRepository;
     Gson gson = new Gson();
